requirements.txt
fastapi
uvicorn
fastapi-users[sqlalchemy2]
fastapi-users[authentication]
fastapi-users[jwt]
sqlalchemy
aiomysql
pydantic
pydantic-settings
python-dotenv
requests